3-8 작성자만 수정삭제 가능

+++ src/main/java/com/board/basic/article/ArticleController.java
@@ -1,13 +1,19 @@
 package com.board.basic.article;
 
+import com.board.basic.user.SiteUser;
+import com.board.basic.user.UserService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.stereotype.Service;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
+import java.security.Principal;
 import java.util.List;
 
 @Controller
@@ -16,6 +22,7 @@ import java.util.List;
 public class ArticleController {
 
     private final ArticleService articleService;
+    private final UserService userService;
 
     @GetMapping("/list")
     public String list(Model model) {
@@ -24,14 +31,17 @@ public class ArticleController {
         return "article_list";
     }
 
+    @PreAuthorize("isAuthenticated()")
     @GetMapping("/create")
     public String create(ArticleForm articleForm) {
         return "article_form";
     }
 
+    @PreAuthorize("isAuthenticated()")
     @PostMapping("/create")
-    public String create(@RequestParam(value = "title") String title, @RequestParam(value = "content") String content) {
-        this.articleService.create(title, content);
+    public String create(@RequestParam(value = "title") String title, @RequestParam(value = "content") String content, Principal principal) {
+        SiteUser siteUser = this.userService.getUser(principal.getName());
+        this.articleService.create(title, content, siteUser);
         return "redirect:/article/list";
     }
 
@@ -42,8 +52,9 @@ public class ArticleController {
         return "article_detail";
     }
 
+    @PreAuthorize("isAuthenticated()")
     @GetMapping("/modify/{id}")
-    public String articleModify(ArticleForm articleForm, @PathVariable("id")Integer id){
+    public String articleModify(ArticleForm articleForm, @PathVariable("id")Integer id,Principal principal){
 
         Article article = this.articleService.getArticle(id);
         articleForm.setTitle(article.getTitle());
@@ -51,19 +62,28 @@ public class ArticleController {
         return "article_form";
     }
 
+    @PreAuthorize("isAuthenticated()")
     @PostMapping("/modify/{id}")
-    public String articleModify(@Valid ArticleForm articleForm, BindingResult bindingResult, @PathVariable("id") Integer id) {
+    public String articleModify(@Valid ArticleForm articleForm, BindingResult bindingResult, @PathVariable("id") Integer id,Principal principal) {
         if (bindingResult.hasErrors()) {
             return "article_form";
         }
         Article article = this.articleService.getArticle(id);
+        if (!article.getAuthor().getUsername().equals(principal.getName())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "수정권한이 없습니다.");
+        }
+
         this.articleService.modify(article, articleForm.getTitle(), articleForm.getContent());
         return String.format("redirect:/article/detail/%s", id);
     }
 
+    @PreAuthorize("isAuthenticated()")
     @GetMapping("/delete/{id}")
-    public String articleDelete(@PathVariable("id")Integer id){
+    public String articleDelete(@PathVariable("id")Integer id,Principal principal){
         Article article = this.articleService.getArticle(id);
+        if (!article.getAuthor().getUsername().equals(principal.getName())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "삭제권한이 없습니다.");
+        }
         this.articleService.delete(article);
         return "redirect:/";
     }
